+++ policyengine_canada/parameters/gov/provinces/yt/tax/income/credits/basic_personal_amount/base_amount.yaml
@@ -1,0 +1,18 @@
+description: Yukon credits this basic personal amount, based on the base amount
+values:
+  2022-01-01: 13_521
+metadata:
+  unit: currency-CAD
+  label: Yukon income basic personal base amount
+  reference:
+  - title: Worksheet for the Yukon 2023 Personal Tax Credits Return
+    href: https://www.canada.ca/content/dam/cra-arc/formspubs/pbg/td1yt-ws/td1yt-ws-23e.pdf#page=1
+  # Basic Personal Amount does not appear in the Yukon income tax act
+  - title: Income Tax Act
+    href: https://laws.yukon.ca/cms/images/LEGISLATION/PRINCIPAL/2002/2002-0118/2002-0118.pdf
+  - title: 2023 Yukon Personal Tax Credits Return
+    href: https://www.canada.ca/content/dam/cra-arc/formspubs/pbg/td1yt/td1yt-23e.pdf#page=1
+  - title: 2022 Yukon Personal Tax Credits Return
+    href: https://www.canada.ca/content/dam/cra-arc/formspubs/pbg/td1yt/td1yt-22e.pdf#page=1
+  - title: Worksheet for the Yukon 2022 Personal Tax Credits Return
+    href: https://www.canada.ca/content/dam/cra-arc/formspubs/pbg/td1yt-ws/td1yt-ws-22e.pdf#page=1
